Moving the fixture
